+++ internal/domain/service/ticket_service.go
@@ -10,6 +10,6 @@ func NewTicketService(repo TicketRepository) *TicketService {
 	return &TicketService{repo: repo}
 }
 
-func (s *TicketService) PurchaseTicket(ctx context.Context, userID int64, from string, to string, purchasePrice float32) (Ticket, error) {
+func (s *TicketService) PurchaseTicket(ctx context.Context, userID uint, from string, to string, purchasePrice float32) (Ticket, error) {
 	return s.repo.Create(ctx, userID, from, to, purchasePrice)
 }
